ocfs2: check the metadate alloc before marking extent written
We need to check the free number of the records in each loop to mark
extent written, because the last extent block may be changed through many
times marking extent written and the 'num_free_extents' also be changed. 
In the worst case, the 'num_free_extents' may become less than the
beginning of the loop.  So we should not estimate the free number of the
records at the beginning of the loop to mark extent written.

Link: http://lkml.kernel.org/r/5A261299.6080403@huawei.com
Signed-off-by: Alex Chen <alex.chen@huawei.com>
Reported-by: John Lightsey <john@nixnuts.net>
Reviewed-by: Jun Piao <piaojun@huawei.com>
Reviewed-by: Joseph Qi <jiangqi903@gmail.com>
Cc: Mark Fasheh <mfasheh@versity.com>
Cc: Joel Becker <jlbec@evilplan.org>
Cc: Junxiao Bi <junxiao.bi@oracle.com>
Cc: Changwei Ge <ge.changwei@h3c.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>