mm, oom: introduce memory.oom_group
The cgroup-aware OOM killer treats leaf memory cgroups as memory
consumption entities and performs the victim selection by comparing them
based on their memory footprint.  Then it kills the biggest task inside
the selected memory cgroup.

But there are workloads, which are not tolerant to a such behavior. 
Killing a random task may leave the workload in a broken state.

To solve this problem, memory.oom_group knob is introduced.  It will
define, whether a memory group should be treated as an indivisible memory
consumer, compared by total memory consumption with other memory consumers
(leaf memory cgroups and other memory cgroups with memory.oom_group set),
and whether all belonging tasks should be killed if the cgroup is
selected.

If set on memcg A, it means that in case of system-wide OOM or memcg-wide
OOM scoped to A or any ancestor cgroup, all tasks, belonging to the
sub-tree of A will be killed.  If OOM event is scoped to a descendant
cgroup (A/B, for example), only tasks in that cgroup can be affected.  OOM
killer will never touch any tasks outside of the scope of the OOM event.

Also, tasks with oom_score_adj set to -1000 will not be killed because
this has been a long established way to protect a particular process from
seeing an unexpected SIGKILL from the OOM killer.  Ignoring this user
defined configuration might lead to data corruptions or other misbehavior.

The default value is 0.

Link: http://lkml.kernel.org/r/20171130152824.1591-5-guro@fb.com
Signed-off-by: Roman Gushchin <guro@fb.com>
Acked-by: Michal Hocko <mhocko@suse.com>
Acked-by: Johannes Weiner <hannes@cmpxchg.org>
Cc: Vladimir Davydov <vdavydov.dev@gmail.com>
Cc: Tetsuo Handa <penguin-kernel@I-love.SAKURA.ne.jp>
Cc: David Rientjes <rientjes@google.com>
Cc: Tejun Heo <tj@kernel.org>
Cc: Michal Hocko <mhocko@kernel.org>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>